Changed villager cargo over to float

+++ Assets/Scripts/RTSCore/Resource.cs
@@ -5,22 +5,22 @@ using Swordfish.Navigation;
 public class Resource : Obstacle
 {
     public ResourceGatheringType type = ResourceGatheringType.None;
-    public int amount = 1000;
+    public float amount = 1000;
 
-    public int GetRemoveAmount(int count)
+    public float GetRemoveAmount(float count)
     {
-        int value = amount - count;
-        int overflow = value < 0 ? Mathf.Abs(value) : 0;
+        float value = amount - count;
+        float overflow = value < 0 ? Mathf.Abs(value) : 0;
 
         return count - overflow;
     }
 
     //  Removes count and returns how much was removed
-    public int TryRemove(int count)
+    public float TryRemove(float count)
     {
         amount -= count;
-
-        int overflow = amount < 0 ? Mathf.Abs(amount) : 0;
+ 
+        float overflow = amount < 0 ? Mathf.Abs(amount) : 0;
 
         if (amount <= 0)
         {
